@@ -4,16 +4,16 @@ import com.asalavei.currencyexchange.api.dto.Currency;
 import com.asalavei.currencyexchange.api.json.JsonCurrency;
 
 import java.util.Collection;
-import java.util.stream.Collectors;
 
 /**
  * Converter for converting between {@link JsonCurrency} and {@link Currency}
  */
 public class JsonCurrencyConverter implements JsonDtoConverter<JsonCurrency, Currency> {
+
     /**
      * Converts the {@link JsonCurrency jsonDto} to the {@link Currency}
      *
-     * @param jsonDto incoming the {@link JsonCurrency jsonDto}
+     * @param jsonDto incoming the {@link JsonCurrency} to be converted
      * @return the converted {@link Currency}
      */
     @Override
@@ -29,7 +29,7 @@ public class JsonCurrencyConverter implements JsonDtoConverter<JsonCurrency, Cur
     /**
      * Converts the {@link Currency dto} to the {@link JsonCurrency}
      *
-     * @param dto incoming the {@link Currency dto}
+     * @param dto incoming the {@link Currency} to be converted
      * @return the converted {@link JsonCurrency}
      */
     @Override
@@ -42,9 +42,17 @@ public class JsonCurrencyConverter implements JsonDtoConverter<JsonCurrency, Cur
                 .build();
     }
 
+    /**
+     * Converts the collection of the {@link Currency dtoCollection} to the collection of the {@link JsonCurrency}
+     *
+     * @param dtoCollection the collection of the {@link Currency} to be converted
+     * @return the converted collection of the {@link JsonCurrency}
+     */
     @Override
     public Collection<JsonCurrency> toJsonDto(Collection<Currency> dtoCollection) {
-        return dtoCollection.stream().map(this::toJsonDto).collect(Collectors.toList());
+        return dtoCollection.stream()
+                .map(this::toJsonDto)
+                .toList();
 
     }
 }
