@@ -2,8 +2,11 @@ package com.asalavei.currencyexchange.api.dbaccess.repositories;
 
 import com.asalavei.currencyexchange.api.dbaccess.entities.EntityCurrency;
 
-public interface CurrencyRepository extends Repository<EntityCurrency> {
-    EntityCurrency findByCode(String code);
+import java.util.Optional;
 
-    Integer getIdByCode(String code);
+/**
+ * Repository for accessing currency entities
+ */
+public interface CurrencyRepository extends Repository<EntityCurrency> {
+    Optional<EntityCurrency> findByCode(String code);
 }
