@@ -2,15 +2,22 @@ package com.asalavei.currencyexchange.api.dbaccess.repositories;
 
 import com.asalavei.currencyexchange.api.dbaccess.entities.EntityCurrency;
 import com.asalavei.currencyexchange.api.dbaccess.entities.EntityExchangeRate;
-import com.asalavei.currencyexchange.api.dbaccess.util.ConnectionUtil;
+import com.asalavei.currencyexchange.api.dbaccess.ConnectionManager;
 import com.asalavei.currencyexchange.api.exceptions.*;
 
-import java.math.BigDecimal;
-import java.sql.*;
-import java.util.ArrayList;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.Collection;
+import java.util.Optional;
 
-public class JdbcExchangeRateDao implements ExchangeRateRepository {
+/**
+ * JDBC implementation of the {@link ExchangeRateRepository}
+ */
+public class JdbcExchangeRateDao extends BaseJdbcDao<EntityExchangeRate> implements ExchangeRateRepository {
+
+    private static final String EXCHANGE_RATE = "%s/%s exchange rate";
 
     @Override
     public EntityExchangeRate save(EntityExchangeRate entity) {
@@ -22,31 +29,13 @@ public class JdbcExchangeRateDao implements ExchangeRateRepository {
                        "WHERE bc.code = ? " +
                        "RETURNING id, base_currency_id, target_currency_id, rate) " +
                        "SELECT i.id AS exchange_rate_id, i.rate AS rate, " +
-                       "bc.id AS base_currency_id, bc.full_name AS base_currency_name, bc.code AS base_currency_code,  bc.sign AS base_currency_sign, " +
+                       "bc.id AS base_currency_id, bc.full_name AS base_currency_name, bc.code AS base_currency_code, bc.sign AS base_currency_sign, " +
                        "tc.id AS target_currency_id, tc.full_name AS target_currency_name, tc.code AS target_currency_code, tc.sign AS target_currency_sign " +
                        "FROM inserted i " +
                        "JOIN currencies bc ON i.base_currency_id = bc.id " +
-                       "JOIN currencies tc ON i.target_currency_id = tc.id;";
+                       "JOIN currencies tc ON i.target_currency_id = tc.id";
 
-        try (Connection connection = ConnectionUtil.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.setBigDecimal(1, entity.getRate());
-            preparedStatement.setString(2, entity.getBaseCurrency().getCode());
-            preparedStatement.setString(3, entity.getTargetCurrency().getCode());
-
-            ResultSet resultSet = preparedStatement.executeQuery();
-
-            if (resultSet.next()) {
-                return getEntityExchangeRate(resultSet);
-            } else {
-                throw new CEDatabaseUnavailableException("Failed to retrieve generated ID.");
-            }
-        } catch (SQLException e) {
-            if (e.getSQLState().startsWith("23")) {
-                throw new CEAlreadyExists("Exchange rate for this currency pair already exists");
-            }
-            throw new CEDatabaseUnavailableException(ExceptionMessages.DATABASE_OPERATION_FAILED, e);
-        }
+        return save(query, entity.getRate(), entity.getTargetCurrency().getCode(), entity.getBaseCurrency().getCode());
     }
 
     @Override
@@ -56,25 +45,15 @@ public class JdbcExchangeRateDao implements ExchangeRateRepository {
                        "tc.id AS target_currency_id, tc.full_name AS target_currency_name, tc.code AS target_currency_code, tc.sign AS target_currency_sign " +
                        "FROM exchange_rates er " +
                        "JOIN currencies bc ON er.base_currency_id = bc.id " +
-                       "JOIN currencies tc ON er.target_currency_id = tc.id";
+                       "JOIN currencies tc ON er.target_currency_id = tc.id " +
+                       "ORDER BY er.id";
 
-        try (Connection connection = ConnectionUtil.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query);
-             ResultSet resultSet = preparedStatement.executeQuery()) {
-            Collection<EntityExchangeRate> exchangeRates = new ArrayList<>();
+        return findAll(query);
 
-            while (resultSet.next()) {
-                exchangeRates.add(getEntityExchangeRate(resultSet));
-            }
-
-            return exchangeRates;
-        } catch (SQLException e) {
-            throw new CEDatabaseUnavailableException(ExceptionMessages.DATABASE_OPERATION_FAILED, e);
-        }
     }
 
     @Override
-    public EntityExchangeRate findByCurrencyPairCodes(String baseCurrencyCode, String targetCurrencyCode) {
+    public Optional<EntityExchangeRate> findByCurrencyCodes(String baseCurrencyCode, String targetCurrencyCode) {
         String query = "SELECT er.id AS exchange_rate_id, er.rate AS rate, " +
                        "bc.id AS base_currency_id, bc.full_name AS base_currency_name, bc.code AS base_currency_code, bc.sign AS base_currency_sign, " +
                        "tc.id AS target_currency_id, tc.full_name AS target_currency_name, tc.code AS target_currency_code, tc.sign AS target_currency_sign " +
@@ -83,45 +62,14 @@ public class JdbcExchangeRateDao implements ExchangeRateRepository {
                        "JOIN currencies tc ON er.target_currency_id = tc.id " +
                        "WHERE (bc.code, tc.code) = (?, ?)";
 
-        try (Connection connection = ConnectionUtil.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.setString(1, baseCurrencyCode);
-            preparedStatement.setString(2, targetCurrencyCode);
-
-            ResultSet resultSet = preparedStatement.executeQuery();
-
-            if (resultSet.next()) {
-                return getEntityExchangeRate(resultSet);
-            } else {
-                throw new CENotFoundException(ExceptionMessages.EXCHANGE_RATE_NOT_FOUND);
-            }
-        } catch (SQLException e) {
-            throw new CEDatabaseUnavailableException(ExceptionMessages.DATABASE_OPERATION_FAILED, e);
-        }
+        return findByCodes(query, baseCurrencyCode, targetCurrencyCode);
     }
 
     @Override
-    public BigDecimal getRateByCurrencyPairIds(Integer baseCurrencyId, Integer targetCurrencyId) {
-        try (Connection connection = ConnectionUtil.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(
-                     "SELECT rate FROM exchange_rates WHERE (base_currency_id, target_currency_id) = (?, ?)")) {
-            preparedStatement.setInt(1, baseCurrencyId);
-            preparedStatement.setInt(2, targetCurrencyId);
+    public EntityExchangeRate update(EntityExchangeRate entity) {
+        String baseCurrencyCode = entity.getBaseCurrency().getCode();
+        String targetCurrencyCode = entity.getTargetCurrency().getCode();
 
-            ResultSet resultSet = preparedStatement.executeQuery();
-
-            if (resultSet.next()) {
-                return resultSet.getBigDecimal("rate");
-            } else {
-                throw new CENotFoundException(ExceptionMessages.EXCHANGE_RATE_NOT_FOUND);
-            }
-        } catch (SQLException e) {
-            throw new CEDatabaseUnavailableException(ExceptionMessages.DATABASE_OPERATION_FAILED, e);
-        }
-    }
-
-    @Override
-    public EntityExchangeRate updateRate(String baseCurrencyCode, String targetCurrencyCode, BigDecimal rate) {
         String query = "UPDATE exchange_rates er " +
                        "SET rate = ? " +
                        "FROM currencies bc, currencies tc " +
@@ -131,25 +79,26 @@ public class JdbcExchangeRateDao implements ExchangeRateRepository {
                        "bc.id AS base_currency_id, bc.full_name AS base_currency_name, bc.code AS base_currency_code, bc.sign AS base_currency_sign, " +
                        "tc.id AS target_currency_id, tc.full_name AS target_currency_name, tc.code AS target_currency_code, tc.sign AS target_currency_sign";
 
-        try (Connection connection = ConnectionUtil.getConnection();
+        try (Connection connection = ConnectionManager.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.setBigDecimal(1, rate);
+            preparedStatement.setBigDecimal(1, entity.getRate());
             preparedStatement.setString(2, baseCurrencyCode);
             preparedStatement.setString(3, targetCurrencyCode);
 
             ResultSet resultSet = preparedStatement.executeQuery();
 
             if (resultSet.next()) {
-                return getEntityExchangeRate(resultSet);
+                return extractEntity(resultSet);
             } else {
-                throw new CENotFoundException(ExceptionMessages.EXCHANGE_RATE_NOT_FOUND);
+                throw createExceptionForEmptyResultSet("update", baseCurrencyCode, targetCurrencyCode);
             }
         } catch (SQLException e) {
-            throw new CEDatabaseUnavailableException(ExceptionMessages.DATABASE_OPERATION_FAILED, e);
+            throw createDatabaseException("update", baseCurrencyCode, targetCurrencyCode);
         }
     }
 
-    private EntityExchangeRate getEntityExchangeRate(ResultSet resultSet) throws SQLException {
+    @Override
+    protected EntityExchangeRate extractEntity(ResultSet resultSet) throws SQLException {
         return EntityExchangeRate.builder()
                 .id(resultSet.getInt("exchange_rate_id"))
                 .baseCurrency(EntityCurrency.builder()
@@ -166,5 +115,38 @@ public class JdbcExchangeRateDao implements ExchangeRateRepository {
                         .build())
                 .rate(resultSet.getBigDecimal("rate"))
                 .build();
+    }
+
+    @Override
+    protected CERuntimeException createAlreadyExistsException(Object... params) {
+        return new CEAlreadyExists(String.format(ExceptionMessage.ALREADY_EXISTS, String.format(EXCHANGE_RATE, params[2], params[1])));
+    }
+
+    @Override
+    protected CERuntimeException createExceptionForEmptyResultSet(String operation, Object... params) {
+        String details;
+
+        switch (operation) {
+            case SAVE_OPERATION -> details = String.format(EXCHANGE_RATE + ": currency not found", params[1], params[2]);
+            case UPDATE_OPERATION -> details = String.format("rate: " + EXCHANGE_RATE + " not found", params[0], params[1]);
+            default -> throw new CEDatabaseException(ExceptionMessage.ERROR_PROCESSING_REQUEST_TO_DATABASE);
+        }
+
+        return new CENotFoundException(String.format(ExceptionMessage.FAILED_OPERATION, operation, details));
+    }
+
+    @Override
+    protected CERuntimeException createDatabaseException(String operation, Object... params) {
+        String details;
+
+        switch (operation) {
+            case SAVE_OPERATION -> details = String.format(EXCHANGE_RATE + " to the database", params[1], params[2]);
+            case READ_OPERATION -> details = String.format(EXCHANGE_RATE + " from the database", params[0], params[1]);
+            case READ_ALL_OPERATION -> details = "exchange rates from the database";
+            case UPDATE_OPERATION -> details = String.format(EXCHANGE_RATE + " in the database", params[0], params[1]);
+            default -> throw new CEDatabaseException(ExceptionMessage.ERROR_PROCESSING_REQUEST_TO_DATABASE);
+        }
+
+        return new CEDatabaseException(String.format(ExceptionMessage.FAILED_OPERATION, operation, details));
     }
 }
